Commit #6
Continued working on the implementation of BFS, trying to fix out the explored attribute so that whenever a cell is added to the queue, it isn't considered explored yet but once all the possible paths that come off of it are added, it is. Also want to make sure the same cell doesn't get added again other than in the context of another cell's path.